+++ .babelrc
@@ -1,6 +1,6 @@
 {
   "ignore": [
     "node_modules",
-    "src/decoder-worker.js"
+    "src/DecoderWorker.js"
   ]
 }
