+++ package.json
@@ -44,7 +44,7 @@
     "build": "npm run build-to5 && npm run build-browser && npm run build-minify",
     "build-browser": "browserify lib/index.js --standalone WavDecoder --outfile build/wav-decoder.js",
     "build-minify": "uglifyjs build/wav-decoder.js -m -r self -o build/wav-decoder.min.js",
-    "build-to5": "babel src --out-dir lib && cp src/decoder-worker.js lib/",
+    "build-to5": "babel src --out-dir lib && cp src/DecoderWorker.js lib/",
     "cover": "babel-node $(npm bin)/isparta cover --report text --report html --report lcov _mocha",
     "lint": "eslint src test",
     "test": "mocha --compilers js:espower-babel/guess",
