+++ src/DecoderWorker.js
